requirements.txt
# Core LangChain Framework (Major version update for stability/features)
langchain==1.0.3
langchain-community==0.4.1
langchain-openai==1.0.1

# Vector Store (Latest stable version of ChromaDB)
chromadb==1.3.0

# Web Interface and Configuration
gradio
python-dotenv==1.0.1
pyyaml==6.0.1